Move minikube build cmd to minikube image build
Use the long name also for examples and usage
