+++ pkg/minikube/cruntime/crio.go
@@ -204,6 +204,7 @@ func (r *CRIO) BuildImage(path string, tag string) error {
 	if tag != "" {
 		args = append(args, "-t", tag)
 	}
+	args = append(args, path)
 	c := exec.Command("sudo", args...)
 	if _, err := r.Runner.RunCmd(c); err != nil {
 		return errors.Wrap(err, "crio build image")
